backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
firebase-admin
pydantic
httpx
websockets
email-validator
python-jose[cryptography]
requests
numpy
scikit-learn
twilio
geopy
